Moved primitative properties into a new material class Optimized shadow rays  (Good SPEEDUP !)

+++ include/ray.hpp
@@ -29,7 +29,7 @@ class Ray {
 
         V3 o; // Offset
         V3 s; // Slope
-        unsigned int depth;
+        short depth;
 };
 
 #endif
